refactor(cli): move the firmware commands out of core/cli.py
Extract `flash`, `verify`, `update` and `restore` into the new
`modules/firmware/cli.py`. The command bodies are moved verbatim; the only
edits are the ones section 3.2 of the plan calls for. `core/cli.py` goes from
1019 to 570 lines.

`flash`, `verify` and `update` were registered implicitly by decorating the
root group with `@cli.command()`. Outside core/cli.py they become plain
`@click.command()` and are registered explicitly in `build_cli()`, like
`restore` already was. The command names the user sees do not change.

The three deferred imports inside the function bodies lose one level of
nesting (`..firmware.X` -> `.X`) and stay deferred: none of `fw_aliases`,
`fw_update` or `restore` is in sys.modules after importing the CLI.

Imports left unused in core/cli.py are dropped: `Flasher`, `run_verification`,
`send_identify_command`, `catnip_get_device`, `time`, and the `print_title` /
`print_subtitle` / `print_alias_item` helpers.

Verified: no diff against tests/snapshots/cli_tree_linux.txt, `catnip flash`
with no argument still resolves its deferred import, and `pytest tests/ -q`
gives 328 passed / 3 failed -- the same preexisting failures.
